pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[project]
name = "bussinbank"
version = "0.1.0"
description = "Your AI personal finance manager – tracks transactions, forecasts cash flow, and gives CFO-level advice."
license = "MIT"
readme = "README.md"
requires-python = ">=3.13,<4.0"
authors = [
    {name = "Sam-oo1", email = "samyakdhole02@gmail.com"}
]
maintainers = [
    {name = "Sam-oo1", email = "samyakdhole02@gmail.com"}
]
keywords = ["finance", "ai", "personal-finance", "langgraph", "pydantic"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "Topic :: Office/Business :: Financial",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Programming Language :: Python :: 3.13"
]
dependencies = [
    "langgraph>=1.0,<2.0",
    "langchain-core>=1.1,<2.0",
    "langchain-community>=0.4,<0.5",
    "langchain-openai>=1.1,<2.0",
    "pydantic>=2.12,<3.0",
    "pandas>=2.2,<3.0",
    "python-dotenv>=1.0,<2.0",
    "rich>=14.2,<15.0",
    "typer[all]>=0.20,<0.21",
    "chromadb>=1.3,<2.0",
    "sentence-transformers>=5.1,<6.0",
    "langchain-groq (>=1.1.0,<2.0.0)",
    "langgraph-prebuilt (==1.0.2)"
]

[project.urls]
Homepage = "https://github.com/yourusername/bussinbank"
Repository = "https://github.com/yourusername/bussinbank"
"Bug Tracker" = "https://github.com/yourusername/bussinbank/issues"

[project.optional-dependencies]
dev = [
    "ruff>=0.6.0,<0.7.0",
    "black>=24.0.0,<25.0.0",
    "pyright>=1.1.0,<2.0.0",
    "pytest>=8.0.0,<9.0.0",
    "pytest-cov>=5.0.0,<6.0.0"
]


[tool.poetry]
package-mode = true

[tool.poetry.dependencies]
python = ">=3.13,<4.0"
langgraph = "^1.0"
langchain-core = "^1.1"
langchain-community = "^0.4"
langchain-openai = "^1.1"
pydantic = "^2.12"
pandas = "^2.2"
python-dotenv = "^1.0"
rich = "^14.2"
typer = {extras = ["all"], version = "^0.20"}
chromadb = "^1.3"
sentence-transformers = "^5.1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.6"
black = "^24.0"
pyright = "^1.1"
pytest = "^8.0"
pytest-cov = "^5.0"